use pages instead of app layout
